Change the fully qualified path in CLibreSSL/module.modulemap to a relative path.
This removes the reference to my own home directory, but unfortunately
still needs to assume that this repo is directly inside the home directory.
This seems to be a limitation of modulemaps, that they have no way of
referring to variable-based build paths.

Fix the fully qualified path in the pbxproj too.